+++ index.html
@@ -50,6 +50,6 @@
     </div>
   </section>
 </div>
-<script src="js/index.js"></script>
+<script type="module" src="js/index.js" ></script>
 </body>
 </html>
